test: split compaction providers out of runtime providers
